requirements.txt
flask==2.3.3
boto3==1.34.34
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
cryptography==41.0.7
gunicorn==21.2.0
pydantic==2.5.2 